+++ frontend/src/metabase/modes/components/drill/SortAction.jsx
@@ -1,55 +1,35 @@
 import { t } from "ttag";
-import { TYPE, isa } from "metabase/lib/types";
-import Dimension from "metabase-lib/lib/Dimension";
+import {
+  sortDrill,
+  sortDrillQuestion,
+} from "metabase-lib/lib/queries/drills/sort-drill";
 
-const INVALID_TYPES = [TYPE.Structured];
+const ACTIONS = {
+  asc: {
+    name: "sort-ascending",
+    section: "sort",
+    buttonType: "sort",
+    icon: "arrow_up",
+    tooltip: t`Sort ascending`,
+  },
+  desc: {
+    name: "sort-descending",
+    section: "sort",
+    buttonType: "sort",
+    icon: "arrow_down",
+    tooltip: t`Sort descending`,
+  },
+};
 
 export default ({ question, clicked }) => {
-  const query = question.query();
-  if (!question.isStructured() || !query.isEditable()) {
+  const drill = sortDrill({ question, clicked });
+  if (!drill) {
     return [];
   }
 
-  if (
-    !clicked ||
-    !clicked.column ||
-    clicked.value !== undefined ||
-    INVALID_TYPES.some(type => isa(clicked.column.base_type, type)) ||
-    !clicked.column.source
-  ) {
-    return [];
-  }
-  const { column } = clicked;
-
-  const fieldRef = query.fieldReferenceForColumn(column);
-  if (!fieldRef) {
-    return [];
-  }
-
-  const [sortDirection, sortFieldRef] = query.sorts()[0] || [];
-  const isAlreadySorted =
-    sortFieldRef != null && Dimension.isEqual(fieldRef, sortFieldRef);
-
-  const actions = [];
-  if (!isAlreadySorted || sortDirection === "desc") {
-    actions.push({
-      name: "sort-ascending",
-      section: "sort",
-      buttonType: "sort",
-      icon: "arrow_up",
-      tooltip: t`Sort ascending`,
-      question: () => query.replaceSort(["asc", fieldRef]).question(),
-    });
-  }
-  if (!isAlreadySorted || sortDirection === "asc") {
-    actions.push({
-      name: "sort-descending",
-      section: "sort",
-      buttonType: "sort",
-      icon: "arrow_down",
-      tooltip: t`Sort descending`,
-      question: () => query.replaceSort(["desc", fieldRef]).question(),
-    });
-  }
-  return actions;
+  const { sortDirections } = drill;
+  return sortDirections.map(sortDirection => ({
+    ...ACTIONS[sortDirection],
+    question: () => sortDrillQuestion({ question, clicked, sortDirection }),
+  }));
 };
